+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/AbstractEngineNewPayloadTest.java
@@ -20,6 +20,7 @@ import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.Executi
 import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.ExecutionEngineJsonRpcMethod.EngineStatus.SYNCING;
 import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.ExecutionEngineJsonRpcMethod.EngineStatus.VALID;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -30,7 +31,7 @@ import static org.mockito.Mockito.when;
 
 import org.hyperledger.besu.consensus.merge.MergeContext;
 import org.hyperledger.besu.consensus.merge.blockcreation.MergeMiningCoordinator;
-import org.hyperledger.besu.datatypes.DataGas;
+import org.hyperledger.besu.datatypes.BlobGas;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.datatypes.VersionedHash;
 import org.hyperledger.besu.datatypes.Wei;
@@ -59,9 +60,11 @@ import org.hyperledger.besu.ethereum.core.Deposit;
 import org.hyperledger.besu.ethereum.core.Withdrawal;
 import org.hyperledger.besu.ethereum.eth.manager.EthPeers;
 import org.hyperledger.besu.ethereum.mainnet.BodyValidation;
+import org.hyperledger.besu.ethereum.mainnet.DefaultProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.DepositsValidator;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSpec;
+import org.hyperledger.besu.ethereum.mainnet.ScheduledProtocolSpec;
 import org.hyperledger.besu.ethereum.mainnet.WithdrawalsValidator;
 import org.hyperledger.besu.ethereum.trie.MerkleTrieException;
 import org.hyperledger.besu.plugin.services.exception.StorageException;
@@ -70,12 +73,14 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Predicate;
 
 import io.vertx.core.Vertx;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentMatcher;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -83,11 +88,16 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @ExtendWith(MockitoExtension.class)
 public abstract class AbstractEngineNewPayloadTest {
 
-  protected final long LONDON_TIMESTAMP = 0;
-  protected final long PARIS_TIMESTAMP = 10;
-  protected final long SHANGHAI_TIMESTAMP = 20;
-  protected final long CANCUN_TIMESTAMP = 30;
-  protected final long EXPERIMENTAL_TIMESTAMP = 40;
+  protected final ScheduledProtocolSpec.Hardfork londonHardfork =
+      new ScheduledProtocolSpec.Hardfork("London", 0);
+  protected final ScheduledProtocolSpec.Hardfork parisHardfork =
+      new ScheduledProtocolSpec.Hardfork("Paris", 10);
+  protected final ScheduledProtocolSpec.Hardfork shanghaiHardfork =
+      new ScheduledProtocolSpec.Hardfork("Shanghai", 20);
+  protected final ScheduledProtocolSpec.Hardfork cancunHardfork =
+      new ScheduledProtocolSpec.Hardfork("Cancun", 30);
+  protected final ScheduledProtocolSpec.Hardfork experimentalHardfork =
+      new ScheduledProtocolSpec.Hardfork("Experimental", 40);
 
   @FunctionalInterface
   interface MethodFactory {
@@ -100,6 +110,24 @@ public abstract class AbstractEngineNewPayloadTest {
         final EngineCallListener engineCallListener);
   }
 
+  static class HardforkMatcher implements ArgumentMatcher<Predicate<ScheduledProtocolSpec>> {
+    private final ScheduledProtocolSpec.Hardfork fork;
+    private final ScheduledProtocolSpec spec;
+
+    public HardforkMatcher(final ScheduledProtocolSpec.Hardfork hardfork) {
+      this.fork = hardfork;
+      this.spec = mock(ScheduledProtocolSpec.class);
+      when(spec.fork()).thenReturn(fork);
+    }
+
+    @Override
+    public boolean matches(final Predicate<ScheduledProtocolSpec> value) {
+      if (value == null) {
+        return false;
+      }
+      return value.test(spec);
+    }
+  }
   // private final MethodFactory methodFactory;
   protected AbstractEngineNewPayload method;
 
@@ -109,7 +137,7 @@ public abstract class AbstractEngineNewPayloadTest {
   protected static final Hash mockHash = Hash.hash(Bytes32.fromHexStringLenient("0x1337deadbeef"));
 
   @Mock protected ProtocolSpec protocolSpec;
-  @Mock protected ProtocolSchedule protocolSchedule;
+  @Mock protected DefaultProtocolSchedule protocolSchedule;
   @Mock protected ProtocolContext protocolContext;
 
   @Mock protected MergeContext mergeContext;
@@ -134,6 +162,12 @@ public abstract class AbstractEngineNewPayloadTest {
         .thenReturn(new DepositsValidator.ProhibitedDeposits());
     lenient().when(protocolSchedule.getByBlockHeader(any())).thenReturn(protocolSpec);
     lenient().when(ethPeers.peerCount()).thenReturn(1);
+    lenient()
+        .when(protocolSchedule.hardforkFor(argThat(new HardforkMatcher(cancunHardfork))))
+        .thenReturn(Optional.of(cancunHardfork));
+    lenient()
+        .when(protocolSchedule.hardforkFor(argThat(new HardforkMatcher(shanghaiHardfork))))
+        .thenReturn(Optional.of(shanghaiHardfork));
   }
 
   @Test
@@ -146,7 +180,9 @@ public abstract class AbstractEngineNewPayloadTest {
             new BlockProcessingResult(Optional.of(new BlockProcessingOutputs(null, List.of()))),
             Optional.empty(),
             Optional.empty());
-
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     assertValidResponse(mockHeader, resp);
@@ -157,7 +193,9 @@ public abstract class AbstractEngineNewPayloadTest {
     BlockHeader mockHeader =
         setupValidPayload(
             new BlockProcessingResult("error 42"), Optional.empty(), Optional.empty());
-
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     EnginePayloadStatusResult res = fromSuccessResp(resp);
@@ -192,7 +230,7 @@ public abstract class AbstractEngineNewPayloadTest {
 
   @Test
   public void shouldReturnSuccessOnAlreadyPresent() {
-    BlockHeader mockHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader mockHeader = createBlockHeader(Optional.empty(), Optional.empty());
     Block mockBlock =
         new Block(mockHeader, new BlockBody(Collections.emptyList(), Collections.emptyList()));
 
@@ -228,7 +266,9 @@ public abstract class AbstractEngineNewPayloadTest {
             new BlockProcessingResult(Optional.empty(), new StorageException("database bedlam")),
             Optional.empty(),
             Optional.empty());
-
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     fromErrorResp(resp);
@@ -245,6 +285,9 @@ public abstract class AbstractEngineNewPayloadTest {
             Optional.empty(),
             Optional.empty());
 
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     verify(engineCallListener, times(1)).executionEngineCalled();
@@ -278,21 +321,27 @@ public abstract class AbstractEngineNewPayloadTest {
 
   @Test
   public void shouldReturnInvalidBlockHashOnBadHashParameter() {
-    BlockHeader mockHeader = new BlockHeaderTestFixture().buildHeader();
+    BlockHeader mockHeader = spy(createBlockHeader(Optional.empty(), Optional.empty()));
     lenient()
         .when(mergeCoordinator.getLatestValidAncestor(mockHeader.getBlockHash()))
         .thenReturn(Optional.empty());
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
+    lenient()
+        .when(mergeCoordinator.latestValidAncestorDescendsFromTerminal(any(BlockHeader.class)))
+        .thenReturn(true);
+    lenient().when(mockHeader.getHash()).thenReturn(Hash.fromHexStringLenient("0x1337"));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     EnginePayloadStatusResult res = fromSuccessResp(resp);
-    assertThat(res.getLatestValidHash()).isEmpty();
     assertThat(res.getStatusAsString()).isEqualTo(getExpectedInvalidBlockHashStatus().name());
     verify(engineCallListener, times(1)).executionEngineCalled();
   }
 
   @Test
   public void shouldCheckBlockValidityBeforeCheckingByHashForExisting() {
-    BlockHeader realHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader realHeader = createBlockHeader(Optional.empty(), Optional.empty());
     BlockHeader paramHeader = spy(realHeader);
     when(paramHeader.getHash()).thenReturn(Hash.fromHexStringLenient("0x1337"));
 
@@ -306,7 +355,7 @@ public abstract class AbstractEngineNewPayloadTest {
 
   @Test
   public void shouldReturnInvalidOnMalformedTransactions() {
-    BlockHeader mockHeader = new BlockHeaderTestFixture().buildHeader();
+    BlockHeader mockHeader = createBlockHeader(Optional.empty(), Optional.empty());
     when(mergeCoordinator.getLatestValidAncestor(any(Hash.class)))
         .thenReturn(Optional.of(mockHash));
 
@@ -320,30 +369,8 @@ public abstract class AbstractEngineNewPayloadTest {
   }
 
   @Test
-  public void shouldReturnInvalidOnOldTimestampInBlock() {
-    BlockHeader parent = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
-    BlockHeader mockHeader =
-        new BlockHeaderTestFixture()
-            .baseFeePerGas(Wei.ONE)
-            .parentHash(parent.getHash())
-            .timestamp(parent.getTimestamp())
-            .buildHeader();
-    when(blockchain.getBlockByHash(mockHeader.getHash())).thenReturn(Optional.empty());
-    when(blockchain.getBlockHeader(parent.getHash())).thenReturn(Optional.of(parent));
-    var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
-
-    assertThat(resp.getType()).isEqualTo(JsonRpcResponseType.SUCCESS);
-    var res = ((JsonRpcSuccessResponse) resp).getResult();
-    assertThat(res).isInstanceOf(EnginePayloadStatusResult.class);
-    var payloadStatusResult = (EnginePayloadStatusResult) res;
-    assertThat(payloadStatusResult.getStatus()).isEqualTo(INVALID);
-    assertThat(payloadStatusResult.getError()).isEqualTo("block timestamp not greater than parent");
-    verify(engineCallListener, times(1)).executionEngineCalled();
-  }
-
-  @Test
   public void shouldRespondWithSyncingDuringForwardSync() {
-    BlockHeader mockHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader mockHeader = createBlockHeader(Optional.empty(), Optional.empty());
     when(mergeContext.isSyncing()).thenReturn(Boolean.TRUE);
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
@@ -356,7 +383,7 @@ public abstract class AbstractEngineNewPayloadTest {
 
   @Test
   public void shouldRespondWithSyncingDuringBackwardsSync() {
-    BlockHeader mockHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader mockHeader = createBlockHeader(Optional.empty(), Optional.empty());
     when(mergeCoordinator.appendNewPayloadToSync(any()))
         .thenReturn(CompletableFuture.completedFuture(null));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
@@ -370,7 +397,7 @@ public abstract class AbstractEngineNewPayloadTest {
 
   @Test
   public void shouldRespondWithInvalidIfExtraDataIsNull() {
-    BlockHeader realHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader realHeader = createBlockHeader(Optional.empty(), Optional.empty());
     BlockHeader paramHeader = spy(realHeader);
     when(paramHeader.getHash()).thenReturn(Hash.fromHexStringLenient("0x1337"));
     when(paramHeader.getExtraData().toHexString()).thenReturn(null);
@@ -387,9 +414,10 @@ public abstract class AbstractEngineNewPayloadTest {
   @Test
   public void shouldReturnInvalidWhenBadBlock() {
     when(mergeCoordinator.isBadBlock(any(Hash.class))).thenReturn(true);
-    BlockHeader mockHeader = new BlockHeaderTestFixture().baseFeePerGas(Wei.ONE).buildHeader();
+    BlockHeader mockHeader = createBlockHeader(Optional.empty(), Optional.empty());
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
-
+    when(protocolSpec.getWithdrawalsValidator())
+        .thenReturn(new WithdrawalsValidator.AllowedWithdrawals());
     EnginePayloadStatusResult res = fromSuccessResp(resp);
     assertThat(res.getLatestValidHash()).contains(Hash.ZERO);
     assertThat(res.getStatusAsString()).isEqualTo(INVALID.name());
@@ -405,7 +433,9 @@ public abstract class AbstractEngineNewPayloadTest {
             new BlockProcessingResult(Optional.of(new BlockProcessingOutputs(null, List.of()))),
             Optional.empty(),
             Optional.empty());
-
+    lenient()
+        .when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+        .thenReturn(Optional.of(mock(BlockHeader.class)));
     var resp = resp(mockPayload(mockHeader, Collections.emptyList()));
 
     assertValidResponse(mockHeader, resp);
@@ -456,8 +486,8 @@ public abstract class AbstractEngineNewPayloadTest {
         header.getPrevRandao().map(Bytes32::toHexString).orElse("0x0"),
         txs,
         withdrawals,
-        header.getDataGasUsed().map(UnsignedLongParameter::new).orElse(null),
-        header.getExcessDataGas().map(DataGas::toHexString).orElse(null),
+        header.getBlobGasUsed().map(UnsignedLongParameter::new).orElse(null),
+        header.getExcessBlobGas().map(BlobGas::toHexString).orElse(null),
         versionedHashes,
         deposits);
   }
@@ -469,8 +499,8 @@ public abstract class AbstractEngineNewPayloadTest {
 
     BlockHeader mockHeader = createBlockHeader(maybeWithdrawals, maybeDeposits);
     when(blockchain.getBlockByHash(mockHeader.getHash())).thenReturn(Optional.empty());
-    when(blockchain.getBlockHeader(mockHeader.getParentHash()))
-        .thenReturn(Optional.of(mock(BlockHeader.class)));
+    // when(blockchain.getBlockHeader(mockHeader.getParentHash()))
+    //  .thenReturn(Optional.of(mock(BlockHeader.class)));
     when(mergeCoordinator.getLatestValidAncestor(any(BlockHeader.class)))
         .thenReturn(Optional.of(mockHash));
     if (validateTerminalPoWBlock()) {